Crate dxgi_win

FFI bindings to DXGI

Reexports

pub use constants::*;

pub use enumerations::*;

pub use structures::*;

pub use interfaces::*;

Modules

constants
Constants and related typedefs provided by DXGI

enumerations
Enumerations provided by DXGI.

interfaces
Interfaces provided by DXGI.

structures
Structures provided by DXGI.

Macros

c_vtable!
Macro used to make making bindings for C representations of C++ classes more convenient.

c_vtable_main!
Helper macro. Actually construct the traits, vtables, and manage the inheritance

c_vtable_methods_to_trait!
Helper macro. Construct a trait with the methods of the C++ interface. Removes the need to manually call the functions of a C vtable struct when forking with C++ class representations

c_vtable_pre!
Helper macro. Deconstruct some syntax before actually doing anything. Handle the pub keyword for vtable and trait, and "with heirs" phrase

c_vtable_struct!
Helper macro. Construct C struct representing C++ method vtable

define_guid!
Rust equivalent to windows C DEFINE_GUID macro

impl_c_vtable_trait!
Helper macro. Implement the trait representing the class methods for the C vtable

Functions

CreateDXGIFactory

CreateDXGIFactory1

CreateDXGIFactory2

DXGIGetDebugInterface

DXGIGetDebugInterface1